implementacion de toda la logica ok

+++ api/app/domain/model/satellite.go
@@ -2,6 +2,6 @@ package model
 
 type Satellite struct {
 	Name string `json:"name"`
-	Distance float32 `json:"distance"`
+	Distance float64 `json:"distance"`
 	Message []string `json:"message"`
 }
